Updated ch2 code after reviewing solution files

+++ functions/write_file.py
@@ -4,9 +4,17 @@ def write_file(working_directory, file_path, content):
     abs_path = os.path.abspath(os.path.join(working_directory, file_path))
     wd_abs_path = os.path.abspath(working_directory)
     if not abs_path.startswith(wd_abs_path):
-        return print(f'Error: Cannot write to "{file_path}" as it is outside the permitted working directory')
+        return f'Error: Cannot write to "{file_path}" as it is outside the permitted working directory'
     if not os.path.exists(file_path):
-        os.makedirs(file_path)
-    with open(abs_path, "w") as f:
-        f.write(content)
-    return print(f'Successfully wrote to "{file_path}" ({len(content)} characters written)')
+        try:
+            os.makedirs(os.path.dirname(abs_path), exist_ok=True)
+        except Exception as e:
+            return f'Error: creating directory {e}'
+    if os.path.exists(abs_path) and os.path.isdir(abs_path):
+        return f'Error: "{file_path}" is a directory, not a file'
+    try: 
+        with open(abs_path, "w") as f:
+            f.write(content)
+        return f'Successfully wrote to "{file_path}" ({len(content)} characters written)'
+    except Exception as e:
+        return f'Error: writing to file: {e}'
